Add example 2 from Rubner, Tomasi, Guibas code as test case for EMD.
Two examples are providd in [1]. This patch adds the test case in
example2.c from [1] as a unittest for Earth Movers Distance (EMD) function
using the original code from [1].

[1] http://robotics.stanford.edu/~rubner/emd/default.htm

Since C++ does not allow nested function definitions and static members for
local classes, the distance function is added as a static method in the
EarthMoverDistanceTest class. Eventually the distance method should be
moved to a functor.