Modify emd() functions to use maximum signature size from the options class.
The Earth Movers Distance (EMD) code imported from sources accompanying the
paper hardcodes the data structure size in with a preprocessor macro.
Switch to using the variable in the options class instead. This requires
dynamic memory allocation, for which we perform no error checking in hopes
that the code will be converted to STL/VIGRA data structures at some point.